Add namespace command
In this commit the all the commands of network a part from 'find' are
moved to the namespace command.
The network avro schema now contains the fields of the 'network find'
result.
