+++ suzieq/config/schema/namespace.avsc
@@ -1,0 +1,82 @@
+{
+    "namespace": "suzieq",
+    "name": "namespace",
+    "type": "derivedRecord",
+    "depends": [
+        "device",
+        "bgp",
+        "ospf",
+        "evpnVni",
+        "mlag",
+        "sqPoller"
+    ],
+    "fields": [
+        {
+            "name": "sqvers",
+            "type": "string",
+            "partition": 0,
+            "default": "2.0",
+            "suppress": true,
+            "description": "Schema version, not selectable"
+        },
+        {
+            "name": "deviceCnt",
+            "type": "long",
+            "display": 1,
+            "description": "Number of devices in inventory in this namespace"
+        },
+        {
+            "name": "serviceCnt",
+            "type": "long",
+            "display": 2,
+            "description": "Number of services polled in this namespace"
+        },
+        {
+            "name": "errSvcCnt",
+            "type": "long",
+            "display": 3,
+            "description": "Number of services with errors in this namespace"
+        },
+        {
+            "name": "hasOspf",
+            "type": "boolean",
+            "display": 4,
+            "description": "Is OSPF used in this namespace"
+        },
+        {
+            "name": "hasBgp",
+            "type": "boolean",
+            "display": 5,
+            "description": "Is BGP used in this namespace"
+        },
+        {
+            "name": "hasVxlan",
+            "type": "boolean",
+            "display": 6,
+            "description": "Is VXLAN used in this namespace"
+        },
+        {
+            "name": "hasMlag",
+            "type": "boolean",
+            "display": 7,
+            "description": "Is MLAG used in this namespace"
+        },
+        {
+            "name": "namespace",
+            "type": "string",
+            "key": 0,
+            "display": 0,
+            "description": "Namespace associated with this record"
+        },
+        {
+            "name": "lastUpdate",
+            "type": "timestamp",
+            "display": 8,
+            "description": "Time when any service was last polled"
+        },
+        {
+            "name": "active",
+            "type": "boolean"
+        }
+    ]
+}
