+++ suzieq/sqobjects/network.py
@@ -12,11 +12,8 @@ class NetworkObj(SqObject):
 
     def __init__(self, **kwargs):
         super().__init__(table='network', **kwargs)
-        self._valid_get_args = ['namespace', 'hostname', 'version', 'os',
-                                'model', 'vendor', 'columns', 'query_str']
         self._valid_find_args = ['namespace', 'hostname', 'vrf', 'vlan',
                                  'address', 'query_str']
-        self._unique_def_column = ['namespace']
 
     def find(self, **kwargs) -> pd.DataFrame():
         '''Find network attach point for a given address'''
